Add perf benchmarks for the ArrowStreamReader.
